requirements.txt
requests~=2.23.0
pillow~=8.4.0
transformers~=4.10.1
torch